requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
coveralls==2.1.2
dataclasses==0.6
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-import-export==2.4.0
django-tables2==2.3.2
docopt==0.6.2
et-xmlfile==1.0.1
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.5
pathspec==0.8.0
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
